Resort FeatureDefinesCairo.vcprops to actually be alphabetical (in preparation for autogeneration) https://bugs.webkit.org/show_bug.cgi?id=85454
Reviewed by Adam Barth.

The PreprocessorDefinitions line is still not sorted, but I'll do that in a separate patch.
There is no functional change in this patch.  Just moving the defines around.
Its interesting to me that now that it's sorted and I can easily compare this with other
ports and found that a whole bunch of defines which should be on for Cairo are off,
likely due to Cairo simply being overlooked.  These oversights should be easy to avoid
in an autogenerated world.

* win/tools/vsprops/FeatureDefinesCairo.vsprops:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115938 268f45cc-cd09-0410-ab3c-d52691b4dbfc