[TextureMapper] Tiles are not created for large textures https://bugs.webkit.org/show_bug.cgi?id=86245
Patch by Martin Robinson <mrobinson@igalia.com> on 2012-05-11
Reviewed by Noam Rosenthal.

No new tests. This will not produce any observable behavior changes,
unless run on a machine with a small texture size limit.

The maxTextureSize() method on TextureMapperGL was missing a "const"
keyword, meaning that it was not properly overriding the version in
the abstract base class (TextureMapper). This patch adds the const
modifier and cleans up the list of override methods in the two
TextureMapper sublcasses, adding the OVERRIDE macro for compilers that
support it and removing a couple unused methods.

* platform/graphics/texmap/TextureMapperGL.cpp:
* platform/graphics/texmap/TextureMapperGL.h:
(WebCore::TextureMapperGL::create):
* platform/graphics/texmap/TextureMapperImageBuffer.h:
(TextureMapperImageBuffer):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116834 268f45cc-cd09-0410-ab3c-d52691b4dbfc