Volume slider needs to be displayed below the mute button https://bugs.webkit.org/show_bug.cgi?id=85990
Patch by Victor Carbune <vcarbune@adobe.com> on 2012-05-13
Reviewed by Dimitri Glazkov.

Source/WebCore:

Added back code for rendering offset, but used only when the volume slider
needs to be displayed below the controls.

Test: media/media-volume-slider-rendered-below.html

* css/mediaControlsChromium.css:
(audio::-webkit-media-controls-volume-slider-container, video::-webkit-media-controls-volume-slider-container):
Changed positioning of the slider to absolute, otherwise it is not possible to position it from the layout() method.
* css/mediaControlsQuickTime.css:
Added back the double mute-button and changed the z-index of the original button. When the slider is displayed
the second mute button is actually there.
(audio::-webkit-media-controls-mute-button, video::-webkit-media-controls-mute-button):
(audio::-webkit-media-controls-volume-slider-container, video::-webkit-media-controls-volume-slider-container):
(audio::-webkit-media-controls-volume-slider-mute-button, video::-webkit-media-controls-volume-slider-mute-button):
Added copy.
* html/shadow/MediaControlElements.cpp:
(WebCore::MediaControlPanelElement::makeTransparent): Disabled the possible of hiding controls. WebVTT rendering
in the place of controls when these are visible is not possible with the current code.
(RenderMediaVolumeSliderContainer):
(WebCore):
(WebCore::RenderMediaVolumeSliderContainer::RenderMediaVolumeSliderContainer):
(WebCore::RenderMediaVolumeSliderContainer::layout):
Checked if the absolute coordinates of the corner of the slider would be rendered outside the page. This part of the
code is faulty if display:none is toggled on the controls.
(WebCore::MediaControlVolumeSliderContainerElement::createRenderer):
* html/shadow/MediaControlElements.h:
(MediaControlVolumeSliderContainerElement):
* html/shadow/MediaControlRootElement.cpp:
(WebCore::MediaControlRootElement::create):

LayoutTests:

Updated tests expectation and marked failing tests accordingly.

* media/media-volume-slider-rendered-below.html: Added.
* media/media-volume-slider-rendered-expected.txt: Added.
* platform/chromium/test_expectations.txt: Updated.
* platform/mac/media/audio-controls-rendering-expected.txt: Updated (z-index change).
* platform/mac/media/controls-after-reload-expected.txt: Same.
* platform/mac/media/controls-strict-expected.txt: Same.
* platform/mac/media/video-controls-rendering-expected.txt: Same.
* platform/mac/media/video-display-toggle-expected.txt: Same.
* platform/mac/media/video-playing-and-pause-expected.txt: Same.
* platform/mac/test_expectations.txt: Updated.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116900 268f45cc-cd09-0410-ab3c-d52691b4dbfc