[Qt] Make sure dialogs have their parent and attached properties set before Component.onCompleted
https://bugs.webkit.org/show_bug.cgi?id=86222

The WebView QML api for dialogs is based on delegates for each dialog, which
are implemented as components. Since there's no explicit API to show the dialog,
the natural place to allow the dialog to choose how to display itself is in
Component.onCompleted.

The issue was that we were using QQmlComponent::create() when instantiating the
component, at which point the component did not have a parent or it's attached
WebView property set. That meant that we had to do some ugly workaround in our
test, using a 1ms timer, to get a callback once the parent was there.

We now use the QQmlComponent::beginCreate() -> QQmlComponent::completeCreate()
pair, which allows us to set the relevant properties before triggering the
Component.onCompleted callback. We were already doing this for the context
menus (except for setting the parent, which we did afterwards).

Reviewed by Simon Hausmann.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116936 268f45cc-cd09-0410-ab3c-d52691b4dbfc