Change ascents and descent back to ints for now https://bugs.webkit.org/show_bug.cgi?id=86518
Reviewed by Eric Seidel.

Change ascents and descent calculation back to integers for now as it
looks like we'll need to support both the 1/60 and 1/1 implementations of
FractionalLayoutUnit for the immediate future to ensure that text is
rendered at the same offset and with the height regardless of the
ENABLE_SUBPIXEL_LAYOUT flag.

This ensures that most websites will look the same at 100% zoom
regardless of the flag and also allows us to the same test expectations
for the vast majority of layout tests.

Eventually we want to move those back to FractionalLayoutUnits as that
will result in more correct rendering.

Source/WebCore: 

No new tests, no change in functionality.

* rendering/InlineFlowBox.cpp:
(WebCore::InlineFlowBox::computeLogicalBoxHeights):
* rendering/RootInlineBox.cpp:
(WebCore::setAscentAndDescent):
(WebCore::RootInlineBox::ascentAndDescentForBox):
* rendering/RootInlineBox.h:
(RootInlineBox):

LayoutTests: 

* platform/chromium-win/fast/block/float/overhanging-tall-block-expected.txt:
* platform/mac/fast/block/float/overhanging-tall-block-expected.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117310 268f45cc-cd09-0410-ab3c-d52691b4dbfc