[chromium] Remove ImageDecoderCG.cpp from platform/image-decoders https://bugs.webkit.org/show_bug.cgi?id=86346
Reviewed by Adam Barth.

ImageDecoderCG.cpp was added for use in the Chromium port in r70846. Remove the
implementation now that Chromium uses Skia on Mac OSX by default.

No new tests. No change in behavior.

* WebCore.gypi: Remove ImageDecoderCG.cpp from the gyp project.
* platform/graphics/ImageSource.h: Revert the r70846 changes.
* platform/image-decoders/ImageDecoder.cpp: Remove copyReferenceToBitmapData().
(WebCore::ImageFrame::operator=): Return to using copyBitmapData() (as it was)
given the removal of copyReferenceToBitmapData() above.
* platform/image-decoders/ImageDecoder.h:
(ImageFrame): Remove copyReferenceToBitmapData() declaration.
* platform/image-decoders/cg/ImageDecoderCG.cpp: Removed, and it was the only
decoder with a port-specific implementation of copyReferenceToBitmapData().


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117355 268f45cc-cd09-0410-ab3c-d52691b4dbfc