Animated GIFs in page cache get updated https://bugs.webkit.org/show_bug.cgi?id=86668 <rdar://problem/11395549>
Reviewed by Brady Eidson.

Source/WebCore:

Test: fast/loader/image-in-page-cache.html

* rendering/RenderImage.cpp:
(WebCore::RenderImage::imageChanged): When we are notified by the CachedImage that the image has
changed, we check to see if the document is in the page cache. If so, we should not be updating,
so we bail out early.

LayoutTests:

This test loads pages with subframes that display the same animated gif. When one of
the pages goes into the page cache, we want to make sure that we do not crash as a
side effect of the page's images getting updated.

* fast/loader/image-in-page-cache-expected.txt: Added.
* fast/loader/image-in-page-cache.html: Added.
* fast/loader/resources/animated-image-subframe.html: Added.
* fast/loader/resources/animated-image.html: Added.
* fast/loader/resources/animated-image2.html: Added.
* fast/loader/resources/animated-image3.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117366 268f45cc-cd09-0410-ab3c-d52691b4dbfc