SVGAnimatedPropertyTearOff does not clear a self pointer on deletion https://bugs.webkit.org/show_bug.cgi?id=86119
Reviewed by Nikolas Zimmermann.

Source/WebCore: 

SVGAnimatedPropertyTearOff contains two SVGPropertyTearOff objects
that have a pointer back to the SVGAnimatedPropertyTearOff. JS may
also have a reference to these SVGPropertyTearOff objects. When the
SVGAnimatedPropertyTearOff is deleted, the SVGPropertyTearOff objects
may live on, but the pointer back to the deleted animated property
tear off is left invalid. This patch clears the pointers on destruction
of the SVGAnimatedPropertyTearOff.

Test: svg/custom/bug86119.html

* svg/properties/SVGAnimatedPropertyTearOff.h:
(WebCore::SVGAnimatedPropertyTearOff::~SVGAnimatedPropertyTearOff):
(SVGAnimatedPropertyTearOff):

LayoutTests: 

* svg/custom/bug86119.html: Added.
* svg/custom/bug86119-expected.txt: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117792 268f45cc-cd09-0410-ab3c-d52691b4dbfc