[Gtk][LayoutTests] Repaint the complete WebKitWebView before dumping pixel results https://bugs.webkit.org/show_bug.cgi?id=86284
Reviewed by Martin Robinson.

Source/WebKit/gtk: 

Add an additional method to the ChromeClient through which it is
possible to bypass the painting timer and force an immediate repaint of
the current dirty regions.

Add a new method to DumpRenderTreeSupportGtk class that calls that
method, making it possible to force an instant paint from
DumpRenderTree.

* WebCoreSupport/ChromeClientGtk.cpp:
(WebKit::ChromeClient::ChromeClient):
(WebKit::ChromeClient::paint): Only delay the painting if the paint is
not forced.
(WebKit::ChromeClient::forcePaint):
(WebKit):
* WebCoreSupport/ChromeClientGtk.h:
(ChromeClient):
* WebCoreSupport/DumpRenderTreeSupportGtk.cpp:
(DumpRenderTreeSupportGtk::getFrameChildren): Refactor getFrameChildren
to use the kit() function on the core child frames rather than using
the FrameLoaderClient of that frame, which is what the kit() function does in the first
place. This makes it possible to remove the inclusion of the FrameLoaderClientGtk.h
header which causes symbol name conflicts when included along with ChromeClientGtk.h
that originate in the Xlib headers included through the former header.
(DumpRenderTreeSupportGtk::forceWebViewPaint):
* WebCoreSupport/DumpRenderTreeSupportGtk.h:
(DumpRenderTreeSupportGtk):

Tools: 

Paint the current dirty regions of the web view once more before
dumping pixel results.

* DumpRenderTree/gtk/DumpRenderTree.cpp:
(dump):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@117947 268f45cc-cd09-0410-ab3c-d52691b4dbfc