NULL ptr in WebCore::RenderBlock::layoutRunsAndFloatsInRange https://bugs.webkit.org/show_bug.cgi?id=77786
Reviewed by Ryosuke Niwa.

Source/WebCore:

InlineBidiResolver adds one fake TextRun for isolated inlines in the process of creating the
list of TextRuns to send to the UBA. After the UBA has been run and the TextRuns reordered,
we re-run InlineBidiResolver rooted in the isolate and replace the fake run with those
generated by the subsequent pass by calling the method BidiRunList::replaceRunWithRuns.
This method assumes there are runs to replace the fake run with.

Positioned inline children are ignored when creating TextRuns, so when an isolated inline
has only positioned children we end up with an empty set of runs to pass to replaceRunWithRuns.
Ideally, we'd remove the fake run and not replace it with anything, but BidiRunList keeps
a pointer to the logically last run, which we're unable to easily re-determine after the UBA
has been run. Instead, we leave the fake run in the list and simply avoid calling
replaceRunWithRuns when we don't have any replacement runs.

Test: fast/block/line-layout/crash-in-isolate-with-positioned-child.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::constructBidiRuns):

LayoutTests:

* fast/block/line-layout/crash-in-isolate-with-positioned-child-expected.txt: Added.
* fast/block/line-layout/crash-in-isolate-with-positioned-child.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118114 268f45cc-cd09-0410-ab3c-d52691b4dbfc