[Shadow] drop event is not fired on nodes in Shadow DOM https://bugs.webkit.org/show_bug.cgi?id=85774
Reviewed by Dimitri Glazkov.

Source/WebCore:

EventHander::updateDragAndDrop re-targets an event to a shadow ancestor node,
however it should be done by adjusting event target. So we don't need to have it.

Tests: fast/dom/shadow/drop-event-for-input-in-shadow.html
       fast/dom/shadow/drop-event-in-shadow.html

* page/EventHandler.cpp:
(WebCore::EventHandler::updateDragAndDrop):

LayoutTests:

Confirms drop event is fired on div and input in Shadow DOM.

* fast/dom/shadow/drop-event-for-input-in-shadow-expected.txt: Added.
* fast/dom/shadow/drop-event-for-input-in-shadow.html: Added.
* fast/dom/shadow/drop-event-in-shadow-expected.txt: Added.
* fast/dom/shadow/drop-event-in-shadow.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118298 268f45cc-cd09-0410-ab3c-d52691b4dbfc