Elements in Shadow DOM are not resizable. https://bugs.webkit.org/show_bug.cgi?id=87342
Reviewed by Ryosuke Niwa.

Source/WebCore:

In RenderLayer::resize(), shadowAncestorNode() was used but if the ancestor node is used,
it is not possible to resize elements in Shadow DOM.

The comments said it is necessary for textarea, however actually it is not necessary now.
Existing test (fast/css/resize-corner-tracking) covers it.

Test: fast/dom/shadow/resize-in-shadow-dom.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::resize):

LayoutTests:

* fast/dom/shadow/resize-in-shadow-dom-expected.txt: Added.
* fast/dom/shadow/resize-in-shadow-dom.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118338 268f45cc-cd09-0410-ab3c-d52691b4dbfc