Transformed fixed position layers have an incorrect overlap map entry https://bugs.webkit.org/show_bug.cgi?id=64201
Reviewed by Darin Adler.

Source/WebCore:

Previously, layers that both had a transform and were fixed position
were not considered as being fixed position in RenderGeometryMap or in
RenderBox::mapLocalToContainer (although this case is not incorrect in
the case of painting, so an external caller likely adjusts for this).

Tests: compositing/layer-creation/fixed-position-and-transform.html
       compositing/layer-creation/fixed-position-under-transform.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::mapLocalToContainer):
* rendering/RenderGeometryMap.cpp:
(WebCore::RenderGeometryMap::mapToAbsolute):

LayoutTests:

Add a test for a layer with both fixed position and a transform to
verify that this is considered as fixed position with respect to the
root.

Also, add a test with a child layer with fixed position and a parent
with a transform to verify that the child is not considered as fixed
position with respect to the root.

* compositing/layer-creation/fixed-position-and-transform-expected.png: Added.
* compositing/layer-creation/fixed-position-and-transform-expected.txt: Added.
* compositing/layer-creation/fixed-position-and-transform.html: Added.
* compositing/layer-creation/fixed-position-under-transform-expected.png: Added.
* compositing/layer-creation/fixed-position-under-transform-expected.txt: Added.
* compositing/layer-creation/fixed-position-under-transform.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118957 268f45cc-cd09-0410-ab3c-d52691b4dbfc