Rename offsetTopLeft in RenderBoxModelObject to something better https://bugs.webkit.org/show_bug.cgi?id=85915
Patch by Shezan Baig <shezbaig.wk@gmail.com> on 2012-05-30
Reviewed by Darin Adler.

Renamed offsetTopLeft in RenderBoxModelObject to
adjustedPositionRelativeToOffsetParent, because it returns the given
startPoint after adjusting it to be relative to the top-left corner of
the offsetParent.  The definition of offsetParent itself is non-trivial
and is documented within the body of RenderObject::offsetParent,
therefore I decided to reuse this term, as-is, in the name of this
function.

No new tests; no functional or visible changes.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::offsetLeft):
(WebCore::RenderBox::offsetTop):
* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::adjustedPositionRelativeToOffsetParent):
(WebCore::RenderBoxModelObject::offsetLeft):
(WebCore::RenderBoxModelObject::offsetTop):
* rendering/RenderBoxModelObject.h:
(RenderBoxModelObject):
* rendering/RenderInline.cpp:
(WebCore::RenderInline::offsetLeft):
(WebCore::RenderInline::offsetTop):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118961 268f45cc-cd09-0410-ab3c-d52691b4dbfc