Move the specification of framework and library linking out of WebKit2.xcodeproj and into WebKit2.xcconfig https://bugs.webkit.org/show_bug.cgi?id=87932
Reviewed by Dan Bernstein.

Doing this makes it easier to vary linking based on SDK or other factors.

* Configurations/WebKit2.xcconfig: Create FRAMEWORK_AND_LIBRARY_LDFLAGS,
which specifies the same list of frameworks and libraries that were
previously in WebKit2's 'Link Binary With Libraries' build phase, and
append it to $(OTHER_LDFLAGS).
* WebKit2.xcodeproj/project.pbxproj: Remove all frameworks and
libraries from the 'Link Binary With Libraries' build phase.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119066 268f45cc-cd09-0410-ab3c-d52691b4dbfc