[SOUP] WebProcess crashes when a download is started from an existing ResourceHandle https://bugs.webkit.org/show_bug.cgi?id=87953
Reviewed by Martin Robinson.

The output stream to write the downloaded data is created in the
didReceiveResponse callback of the download client. When a
download is created for an existing ResourceHandle (this happens
for example when policy decision is download), the response has
already been received. In this case we should make sure that the
download client is notified about the response, so that when data
actually arrives the output stream has already been created.

* WebProcess/Downloads/soup/DownloadSoup.cpp:
(WebKit::Download::startWithHandle):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119107 268f45cc-cd09-0410-ab3c-d52691b4dbfc