[Chromium] imageSmoothingEnabled should apply to patterns https://bugs.webkit.org/show_bug.cgi?id=88103
Patch by Keyar Hood <keyar@chromium.org> on 2012-06-06
Reviewed by Stephen White.

Source/WebCore:

Test: fast/canvas/canvas-imageSmoothingEnabled-patterns.html

When drawing patterns on a Canvas, the imageSmoothingEnabled flag is
respected.
* platform/graphics/skia/PlatformContextSkia.cpp:
(WebCore::PlatformContextSkia::setupShader):

LayoutTests:

Added tests for imageSmoothingEnabled flag applying to patterns.

Modified the TestExpectations file as tests are failing due to the new
change. It makes the pictures prettier (and it agrees with other
browsers) so we have decided that the new results are to be considered
correct.
* fast/canvas/canvas-imageSmoothingEnabled-patterns.html: Added.
* platform/chromium-linux/fast/canvas/canvas-imageSmoothingEnabled-patterns-expected.png: Added.
* platform/chromium-linux/fast/canvas/canvas-imageSmoothingEnabled-patterns-expected.txt: Added.
* platform/chromium/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119591 268f45cc-cd09-0410-ab3c-d52691b4dbfc