Incorrect rect-based hit-test result for culled-inline elements https://bugs.webkit.org/show_bug.cgi?id=85849
Patch by Raymes Khoury <raymes@chromium.org> on 2012-06-07
Reviewed by Levi Weintraub.

Source/WebCore:

Modified code which blindly adds culled inlines to rect-based hit-test
results so that it only does so if the child node does not fully cover
the hit-test region.

Test: fast/dom/nodesFromRect-culled-inline.html

* rendering/HitTestResult.cpp:
(WebCore::HitTestResult::addNodeToRectBasedTestResult):

LayoutTests:

Added tests for rect-based hit-testing for the case when the child of a
culled inline element is in the hit-test result. One of these tests is
expected to fail due to https://bugs.webkit.org/show_bug.cgi?id=88376.

* fast/dom/nodesFromRect-culled-inline-expected.txt: Added.
* fast/dom/nodesFromRect-culled-inline.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119733 268f45cc-cd09-0410-ab3c-d52691b4dbfc