[Chromium] features.gypi should have the correct definitions for Android https://bugs.webkit.org/show_bug.cgi?id=88533
Reviewed by Adam Barth.

Make all features which are enabled on Android but disabled on other platforms,
or are disabled on Android while enabled on other platforms, conditional. This
also means that the selection of code compiled for Android will slightly change.

This change also removes the enable_viewport variable as the value of this is
no longer respected by WebKit - it's enabled by default for all of Chromium.
Furthermore, ENABLE_OVERFLOW_SCROLLING is now listed in features.gypi again.

* features.gypi:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119749 268f45cc-cd09-0410-ab3c-d52691b4dbfc