[chromium] Skip willDraw() and didDraw() on fully occluded layers https://bugs.webkit.org/show_bug.cgi?id=88435
Reviewed by Adrienne Walker.

Source/WebCore:

Current willDraw() is called on all layers with non-empty
visibleLayerRect and non-empty scissorRect. This excludes
layers outside the viewport, but does not exclude occluded
layers. We add a check for occlusion to calculateRenderPasses
in order to avoid willDraw() when it will be culled anyway.

We prevent didDraw() from being called for occluded layers, for
which we did not call didDraw() by holding a vector of layers
for which we did call willDraw(). This lets us avoid storing a
flag on the layers, or computing occlusion again in
didDrawAllLayers.

Unit test: CCLayerTreeHostImplTest.willDrawNotCalledOnOccludedLayer

* platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp:
(WebCore::CCLayerTreeHostImpl::calculateRenderPasses):
(WebCore::CCLayerTreeHostImpl::prepareToDraw):
(WebCore::CCLayerTreeHostImpl::didDrawAllLayers):
* platform/graphics/chromium/cc/CCLayerTreeHostImpl.h:
(FrameData):
(CCLayerTreeHostImpl):

Source/WebKit/chromium:

* tests/CCLayerTreeHostImplTest.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119867 268f45cc-cd09-0410-ab3c-d52691b4dbfc