Some additional tweaks to keep TreeShared code simple/clean https://bugs.webkit.org/show_bug.cgi?id=88738
Reviewed by Adam Barth.

* platform/TreeShared.h: Make constructor and destructor both protected.
It's slightly dangerous to have a public non-virtual destructor, and there
is no reason it needs to be public.
(WebCore::TreeShared::~TreeShared): Make debug-only destructor non-virtual;
we were getting no value from it being virtual. Now that it's non-virtual,
there is no need to use an #if on it since it's an empty inline destructor
in non-debug builds.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119943 268f45cc-cd09-0410-ab3c-d52691b4dbfc