[JSC/V8] Improve DOMString[] support https://bugs.webkit.org/show_bug.cgi?id=89070
Reviewed by Kentaro Hara.

Tests: new test case in TestObj.idl

* bindings/js/JSDOMBinding.cpp:
(WebCore::jsArray): Implement jsArray() for DOMStringList.
* bindings/js/JSDOMBinding.h: Declare jsArray() for DOMStringList.
* bindings/scripts/CodeGeneratorJS.pm:
(NativeToJSValue): Avoid to include JSDOMString.h and DOMString.h.
* bindings/scripts/CodeGeneratorV8.pm:
(GetNativeType): Move a check for DOMString[] because it's a RefPtr type
and we'd like to avoid to generate RefPtr<DOMString[]>
(%non_wrapper_types): DOMString[] is not a wrapper type. This is needed
to avoid to generate unnecessary custom function template.
(NativeToJSValue): Avoid to include V8DOMString.h and DOMString.h.

* bindings/scripts/test/TestObj.idl: Add non-overload function using DOMString[].
* bindings/scripts/test/JS/JSTestObj.cpp: Update expectation.
* bindings/scripts/test/JS/JSTestObj.h: ditto.
* bindings/scripts/test/V8/V8TestObj.cpp: ditto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@120304 268f45cc-cd09-0410-ab3c-d52691b4dbfc