On Mac, autocorrection sometimes fails to take place in Safari. https://bugs.webkit.org/show_bug.cgi?id=89982
Reviewed by Darin Adler.

Existing test was turned off due to intermittent failure, which is caused by autocorrection result depending on user data
that may be altered by previous test runs. Hopefully we can turn the test back on once we have a way to make autocorrection
behave consistently.

Basically we should check the value of shouldCheckForCorrection, not shouldShowCorrectionPanel, to determine if we should
early return in markAndReplaceFor().

* editing/Editor.cpp:
(WebCore::Editor::markAndReplaceFor):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121271 268f45cc-cd09-0410-ab3c-d52691b4dbfc