Remove warning about protected values when the Heap is being destroyed https://bugs.webkit.org/show_bug.cgi?id=90302
Reviewed by Geoffrey Garen.

Having to do book-keeping about whether values allocated from a certain
VM are or are not protected makes the JSC API much more difficult to use
correctly. Clients should be able to throw an entire VM away and not have
to worry about unprotecting all of the values that they protected earlier.

* heap/Heap.cpp:
(JSC::Heap::lastChanceToFinalize):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121607 268f45cc-cd09-0410-ab3c-d52691b4dbfc