[Qt] Get rid of un-needed QT += declarative for Qt 5
The declarative module has been renamed to quick1 in Qt 5, and the
engine-only module for Qt 5 is named 'qml'. For many of the instances
we could just remove 'declarative', since the project file was only
used for Qt5/WebKit2 builds. In the other cases the module was wrapped
in a haveQt(4) scope.

Reviewed by Csaba Osztrogonác.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121851 268f45cc-cd09-0410-ab3c-d52691b4dbfc