[Gtk] Add a configuration option for disabling unstable features in releases https://bugs.webkit.org/show_bug.cgi?id=87995
Reviewed by Martin Robinson.

.: 

Add a configuration flag for enabling the unstable features - features of which
support in the Gtk port is being worked on but is not yet complete. The primary
use of this flag is when compiling through the build-webkit script.

All the features that are currently enabled when building through build-webkit but
are disabled by default when executing the configure script directly have their default
value (when the correspondent flag is not passed) set to 'yes' when unstable features
are enabled and 'no' otherwise. This way unstable features are kept disabled when performing
a release build (unless they are specifically enabled).

* configure.ac:

Tools: 

Pass the --enable-unstable-features flag when building the Gtk port through build-webkit.

* Scripts/webkitdirs.pm:
(buildAutotoolsProject):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121988 268f45cc-cd09-0410-ab3c-d52691b4dbfc