Re-enable __declspec(dl{import,export}) for MinGW when using EXPORT_MACROS https://bugs.webkit.org/show_bug.cgi?id=90612
Patch by Jocelyn Turcotte <jocelyn.turcotte@nokia.com> on 2012-07-10
Reviewed by Csaba Osztrogonác.

MinGW was switched to use auto import/export of symbols on r44184.
From my understanding of the documentation, MinGW will not auto-export symbols
unless there are no explicit __declspec(dlexport) in the DLL already.

The issues that originally made us rely on the auto-import feature of MinGW
should now be resolved with the EXPORT_MACROS work. This patch re-enables them.

It also removes the GCC check for internal symbols hiding as the visibility should
already be hidden by default for both MSVC and GCC on Windows anyway.

* wtf/ExportMacros.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122211 268f45cc-cd09-0410-ab3c-d52691b4dbfc