webkitpy: clean up logging handlers, lint common.message_pool https://bugs.webkit.org/show_bug.cgi?id=91152
Reviewed by Ojan Vafai.

The unix implementation of multiprocessing clones any logging
handlers from the parent process into the child; we currently
don't want this behavior in our code, so I was hand-removing the
installed handlers in the child process I knew about. After thinking
about it further, I think it was simpler and safe enough to just
remove all handlers in the child, since the message pool
propagates any message from the child back into the parent.

We can always change this in the future if it turns out to be an issue.

I'm also fixing a couple of other lint warnings while I'm at it.

* Scripts/webkitpy/common/message_pool.py:
(_MessagePool.__exit__):
(_MessagePool._handle_worker_exception):
(_Worker._set_up_logging):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122517 268f45cc-cd09-0410-ab3c-d52691b4dbfc