REGRESSION(r123169): It made fast/js/dfg-inline-arguments-use-from-uninlined-code.html fail on 32 bit platforms https://bugs.webkit.org/show_bug.cgi?id=92002
Source/JavaScriptCore: 

Reviewed by Mark Hahnenberg.
        
In the process of changing the nature of local variable typing, I forgot to modify one of the places where
we glue the DFG's notion of variable prediction to the runtime's notion of variable tagging.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::compile):

LayoutTests: 

Unskip test that this change fixes, as requested by Ossy.

* platform/qt/Skipped:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123393 268f45cc-cd09-0410-ab3c-d52691b4dbfc