Report the extra memory used by immutable StylePropertySet objects. <http://webkit.org/b/92032>
Reviewed by Anders Carlsson.

Add an optional parameter to the MemoryClassInfo constructor for passing in additional base object
size on top of sizeof(T).

Use this in StylePropertySet::reportMemoryUsage() to properly account for the CSSProperty array
tacked onto the end of the object when m_isMutable == false.

* css/StylePropertySet.h:
(WebCore::StylePropertySet::reportMemoryUsage):
* dom/MemoryInstrumentation.h:
(WebCore::MemoryObjectInfo::reportObjectInfo):
(WebCore::MemoryClassInfo::MemoryClassInfo):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123396 268f45cc-cd09-0410-ab3c-d52691b4dbfc