inline-table wrapper should be generated for display: inline element only https://bugs.webkit.org/show_bug.cgi?id=92054
Reviewed by Abhishek Arya.

Source/WebCore:

CSS 2.1 only takes into account the display when determining if we should generate
an inline-table. My misreading of the spec would make us generate an inline-table
wrapper for any inline formatting context.

Test: fast/table/inline-block-generates-table-wrapper.html

* rendering/RenderTable.cpp:
(WebCore::RenderTable::createAnonymousWithParentRenderer):
Changed to use display instead of isInline to properly match the spec.

LayoutTests:

For some reason, using dumpAsText didn't dump the 2 lines properly so the test
is a reftest to work-around that.

* fast/table/inline-block-generates-table-wrapper-expected.html: Added.
* fast/table/inline-block-generates-table-wrapper.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123492 268f45cc-cd09-0410-ab3c-d52691b4dbfc