Implement ColorSuggestionPicker page popup https://bugs.webkit.org/show_bug.cgi?id=92109
Reviewed by Kent Tamura.

.:

* ManualTests/forms/color-suggestion-picker.html: Added. Tests colorSuggestionPicker.{js,css}

Source/WebCore:

Implements ColorSuggestionPicker page popup. This will be used in
certain ports as the datalist UI for <input type=color>.

No new tests because there is no way to open the popup yet.

* Resources/colorSuggestionPicker.css: Added.
(body):
(#main):
(.color-swatch):
(.color-swatch:hover):
(.color-swatch-container):
(.other-color):
* Resources/colorSuggestionPicker.js: Added.
(getScrollbarWidth):When there are more than 20 colors we need to
increase the window width to accommodate the scroll bar.
(createElement):
(handleMessage):
(initialize):
(handleArgumentsTimeout):
(validateArguments):
(submitValue): Calls setValueAndClosePopup with a numValue of 0.
(handleCancel): Calls setValueAndClosePopup with a numValue of -1.
(chooseOtherColor): Calls setValueAndClosePopup with a numValue of -2.
(ColorPicker):
(ColorPicker.prototype._layout): Lays out the color swatches in a 5x4
grid.
(ColorPicker.prototype._handleSwatchClick):
* WebCore.gyp/WebCore.gyp: Added ColorSuggestionPicker action.
* make-file-arrays.py:
(main): Windows gets confused when "&&" is in the command, so we will replace " AND " with " && " internally.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123833 268f45cc-cd09-0410-ab3c-d52691b4dbfc