Blocks with reverse column progression don’t have layout overflow for overflowing columns https://bugs.webkit.org/show_bug.cgi?id=92440
Reviewed by Sam Weinig.

Source/WebCore: 

Test: fast/multicol/progression-reverse-overflow.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::addLayoutOverflow): If column progression is reversed, then allow
layout overflow to go in the opposite direction than it would normally be allowed to go.
For example, in a block with writing-mode: horizontal-tb, direction: ltr and
column-progression: reverse, columns overflow to the left, so layout overflow would go on
the left.

LayoutTests: 

* fast/multicol/progression-reverse-overflow-expected.html: Added.
* fast/multicol/progression-reverse-overflow.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123835 268f45cc-cd09-0410-ab3c-d52691b4dbfc