changing -webkit-order should change the paint order of flex items https://bugs.webkit.org/show_bug.cgi?id=92041
Reviewed by Ojan Vafai.

Source/WebCore:

Override paintChildren and use the flex order iterator to determine the order to paint the children.

Test: css3/flexbox/order-painting.html

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::layoutBlock): Save a reference to the order iterator.
(WebCore::RenderFlexibleBox::paintChildren):
* rendering/RenderFlexibleBox.h:
(RenderFlexibleBox): Hold a reference to the order iterator so we don't have to recreate it at paint time.
    Also mark all the virtual methods with OVERRIDE.

LayoutTests:

Use a ref test since this is testing paint behavior.

* css3/flexbox/order-painting-expected.html: Added.
* css3/flexbox/order-painting.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123842 268f45cc-cd09-0410-ab3c-d52691b4dbfc