Show the unavailable plug-in indicator for Java applets as well https://bugs.webkit.org/show_bug.cgi?id=92521
Reviewed by Sam Weinig.

Source/WebCore:

Now that <applet> behaves more like <embed> and <object>, make sure that we show the unavailable plug-in indicator
and call the correct error callbacks if we fail to instantiate the plug-in.

* WebCore.exp.in:
Export a symbol needed by WebKit2.

* html/HTMLAppletElement.cpp:
(WebCore::HTMLAppletElement::HTMLAppletElement):
Set the correct service type.

* loader/SubframeLoader.cpp:
(WebCore::SubframeLoader::createJavaAppletWidget):
Enable the unavailable plug-in indicator if we fail to create the java applet widget.

Source/WebKit/mac:

Check if the Java plug-in is blocked before trying to instantiate it.

* WebCoreSupport/WebFrameLoaderClient.mm:

Source/WebKit2:

* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::unavailablePluginButtonClicked):
This can now be called on applet elements as well.

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::createJavaAppletWidget):
Get the MIME type from the applet element.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123907 268f45cc-cd09-0410-ab3c-d52691b4dbfc