[Qt][NRWT] REGRESSION(123729): Forcing pixel tests with -p doesn't work https://bugs.webkit.org/show_bug.cgi?id=92627
Patch by Balazs Kelemen <kbalazs@webkit.org> on 2012-07-30
Reviewed by Noam Rosenthal.

Remove the default behavior I added in r123729 so that we run all tests as pixel
test with -p. It was intentional but it seems like we don't have consensus on
that we want it.

* Scripts/webkitpy/layout_tests/port/qt.py:
(QtPort._supports_switching_pixel_tests_per_test):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124020 268f45cc-cd09-0410-ab3c-d52691b4dbfc