Caret position is wrong when a editable container has word-wrap:normal set https://bugs.webkit.org/show_bug.cgi?id=89649
Patch by Pravin D <pravind.2k4@gmail.com> on 2012-07-31
Reviewed by Levi Weintraub.

Source/WebCore:

Correcting the incorrect addumption that the min and max position for the caret rect is constrained to
the logical containing block width. This assumption is incorrect when the text has word-wrap:normal and
flows beyound the visible rect of the container.

Test: editing/input/editable-container-with-word-wrap-normal.html

* rendering/RenderText.cpp:
(WebCore::RenderText::localCaretRect):

LayoutTests:

* editing/input/editable-container-with-word-wrap-normal-expected.txt: Added.
* editing/input/editable-container-with-word-wrap-normal.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124231 268f45cc-cd09-0410-ab3c-d52691b4dbfc