REGRESSION(r120979): Node list returned by getElementsByTagNameNS isn't invalidated when children changes https://bugs.webkit.org/show_bug.cgi?id=92705
Reviewed by Abhishek Arya.

Source/WebCore: 

The bug was caused by a typo in NodeListsNodeData::invalidateCaches. We need to invalidate
tag name node lists when child nodes change and not when attributes change.

Test: fast/dom/NodeList/nodelist-namespace-invalidation.html

* dom/Node.cpp:
(WebCore::NodeListsNodeData::invalidateCaches):

LayoutTests: 

Add a regression test to ensure node lists returned by getElementsByTagNameNS are invalidated
when child nodes are added or removed.

* fast/dom/NodeList/nodelist-namespace-invalidation-expected.txt: Added.
* fast/dom/NodeList/nodelist-namespace-invalidation.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124282 268f45cc-cd09-0410-ab3c-d52691b4dbfc