Move getLocalTime() as static inline function to DateMath https://bugs.webkit.org/show_bug.cgi?id=92955
Reviewed by Ryosuke Niwa.

getCurrentLocalTime() and getLocalTime() has been superseded with the
GregorianDateTime class. So we can move it into DateMath.cpp as an static inline
function. This allows us to remove the dependecy on time() and localtime()
for Windows CE, where this functions require the ce_time library to work.

Source/JavaScriptCore:

* JavaScriptCore.vcproj/JavaScriptCore/JavaScriptCore.def:

Source/WTF:

* wtf/CurrentTime.cpp:
(WTF):
* wtf/CurrentTime.h:
* wtf/DateMath.cpp:
(WTF):
(WTF::getLocalTime):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124560 268f45cc-cd09-0410-ab3c-d52691b4dbfc