Apply target fuzzing when sending a context menu event https://bugs.webkit.org/show_bug.cgi?id=92914
Patch by Terry Anderson <tdanderson@chromium.org> on 2012-08-03
Reviewed by Antonio Gomes.

Source/WebCore:

If TOUCH_ADJUSTMENT is enabled, use bestClickableNodeForTouchPoint to possibly
adjust the location of a context menu event. This change uses the same set of
candidates for touch adjustment as is used for a GestureTap event (which
admittedly is a simplifying assumption).

Test: touchadjustment/touch-links-longpress.html

* page/EventHandler.cpp:
(WebCore::EventHandler::handleGestureTap):
Some code moved to the new function adjustGesturePosition.
(WebCore):
(WebCore::EventHandler::adjustGesturePosition):
Added this function to avoid repeated code in handleGestureTap and
sendContextMenuEventForGesture.
(WebCore::EventHandler::sendContextMenuEventForGesture):
Try to adjust the location of |mouseEvent| to correspond to the most
probable gesture target. If no such target exists, the location of
|mouseEvent| is unchanged.
* page/EventHandler.h:
(EventHandler):

Source/WebKit/chromium:

When constructing a PlatformEvent::GestureLongPress in WebInputEventConversion.cpp,
use |boundingBox| to specify |m_area|.

* src/WebInputEventConversion.cpp:
(WebKit::PlatformGestureEventBuilder::PlatformGestureEventBuilder):

LayoutTests:

New layout test to verify that context menus appear on links for a long press
gesture when TOUCH_ADJUSTMENT is enabled.

* platform/qt/Skipped:
Skipped on qt because gestureLongPress is not used for context menus.
* touchadjustment/touch-links-longpress-expected.txt: Added.
* touchadjustment/touch-links-longpress.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124642 268f45cc-cd09-0410-ab3c-d52691b4dbfc