[chromium] Non-preserves-3d requires explicit flattening of screen-space transform. https://bugs.webkit.org/show_bug.cgi?id=85808
Reviewed by Adrienne Walker.

Source/WebCore:

When computing the screen-space transforms, z values were not
being flattened when they should be. This caused incorrect
clipping due to occlusion tracking that relied on the screen-space
transform. The fix is to flatten the screen-space transform just
like the "sublayerMatrix" is flattened, when the layer does not
preserve-3d.

In addition to making the simple fix, it was convenient to make a
helper function for the flattening code.

Additional unit test added to exercise that flattening code:
  CCLayerTreeHostCommonTest.verifyTransformsForFlatteningLayer

* platform/graphics/chromium/cc/CCLayerTreeHostCommon.cpp:
(WebCore::calculateDrawTransformsInternal):
* platform/graphics/chromium/cc/CCMathUtil.cpp:
(WebCore::CCMathUtil::flattenTransformTo2d):
(WebCore):
* platform/graphics/chromium/cc/CCMathUtil.h:
(CCMathUtil):

Source/WebKit/chromium:

* tests/CCLayerTreeHostCommonTest.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@124782 268f45cc-cd09-0410-ab3c-d52691b4dbfc