showNodePath should show the age of ShadowRoot https://bugs.webkit.org/show_bug.cgi?id=93347
Patch by Takashi Sakamoto <tasak@google.com> on 2012-08-08
Reviewed by Ryosuke Niwa.

No new tests, because this patch is for fixing a bug of a method for
debugging. If NDEBUG is defined, the method is not compiled.

* dom/Node.cpp:
(WebCore::Node::showNodePathForThis):
Fixed the bug to obtain the oldest shadow root from a shadow root.
oldestShadowRootFor only works for an element node, not for a shadow root.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125000 268f45cc-cd09-0410-ab3c-d52691b4dbfc