Crash when inspecting an element with border-image https://bugs.webkit.org/show_bug.cgi?id=93380
Patch by Matt Arsenault <arsenm2@gmail.com> on 2012-08-08
Reviewed by Tim Horton.

Source/WebCore:

The second value in the CSSPair should be the same as the first if
we are in a shorthand and the next value is not a
border-image-repeat keyword.

Test: fast/css/parse-border-image-repeat-null-crash.html

* css/CSSParser.cpp:
(WebCore::CSSParser::parseBorderImageRepeat):

LayoutTests:

Add a test that ensures css Text on a border image shorthand with
border-image-repeat does not crash.

* fast/css/parse-border-image-repeat-null-crash-expected.txt: Added.
* fast/css/parse-border-image-repeat-null-crash.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@125016 268f45cc-cd09-0410-ab3c-d52691b4dbfc